STDAM_generation/requirements.txt
pytorch==1.7.1
sklearn
pandas
numpy
tqdm
scipy

